requirements.txt
dotenv
slack_bolt
requests
osu.py
Flask